asymmetric_mobility_se1e2iiar_2strains_cpp

Raw CPP function. Should not be called directly.

Usage

asymmetric_mobility_se1e2iiar_2strains_cpp(
  se1e2iiar_2strains_pop,
  mobility_matrix,
  seed_matrix,
  seed_matrix_b,
  betas,
  inputSeed,
  a1,
  a2,
  gamma,
  relativeInfectiousnessB,
  presymptomaticRelativeInfectiousness,
  asymptomaticProb,
  asymptomaticRelativeInfectiousness,
  N = 1L,
  M = 56L
)

Arguments

se1e2iiar_2strains_pop

Data.frame

mobility_matrix

List of data.frames

seed_matrix

matrix of seeding cases of strain a per date per geographical location

seed_matrix_b

matrix of seeding cases of strain b per date per geographical location

betas

matrix of floats, number of time intervals times number of locations, infection parameter, 0.6

inputSeed

Integer, input seed

a1

Float, 1/latent period, 1/2.0

a2

Float, 1/presymptomatic period, 1/3.0

gamma

Float, 1/infectious period, 1/5.0

relativeInfectiousnessB

Float, Relative infectiousness of strain B

presymptomaticRelativeInfectiousness

Float, Relative infectiousness of presymptomatic infectious

asymptomaticProb

Float, Proportion/probability of asymptomatic given infectious

asymptomaticRelativeInfectiousness

Float, Relative infectiousness of asymptomatic infectious

N

Int = 1 int, Number of repetitions

M

Int, Number of days