commuter

This model is a stochastic SEIIaR (susceptible, exposed, infectious, infectious asymptomatic, recovered) metapopulation model that including commuting.

Usage

commuter(
  seiiar = spread::nor_seiiar_oslo_2017_b2020,
  commuters = spread::nor_commuters_2017_b2020,
  r0 = NULL,
  beta = NULL,
  latent_period = 1.9,
  infectious_period = 3,
  asymptomatic_prob = 0,
  asymptomatic_relative_infectiousness = 0,
  days_simulation = 7 * 8,
  N = 1,
  simulations = 1,
  verbose = TRUE,
  seed = NULL,
  aggregate_location = FALSE
)

Arguments

seiiar

Data frame containing `location_code`, `S`, `E`, `I`, `Ia`, and `R` for the entire population

commuters

Data frame comtaining `from`, `to`, `n` for the number of people who travel

r0

Float, basic reproduction number

beta

Float, infection parameter, 0.6

latent_period

Float, 1.9

infectious_period

Float, 3

asymptomatic_prob

Float, Proportion/probability of asymptomatic given infectious

asymptomatic_relative_infectiousness

Float, Relative infectiousness of asymptomatic infectious

days_simulation

Int, Number of days to simulate

N

Int = 1 int, Number of internal simulations (average taken). This is generally used for parameter fitting.

simulations

Int, Number of simulations (all returned). This is generally used to generate stochastic results.

verbose

boolean

seed

seed

aggregate_location

Do you want to aggregate over all location_codes?

Value

year

The middle year of a 5 year range (e.g. 2011 is the average of data from 2009-2013).

location_code

The location code.

age

The population age.

vax

The vaccine.

proportion

Proportion of people who are vaccinated.

imputed

FALSE if real data. TRUE if it is the national average.

Details

Each location has a local infection system, while the locations are connected by people who commute each day. The model differentiates between day and night. During the day you can infect/be infected in the location where you work, while during the night you can infect/be infected in the location where you live. It is the same commuters who travel back and forth each day. At the start of a day, all commuters are sent to their work location, where they mix for 12 hours. The commuters are then sent to their respective home locations, where they mix for 12 hours.

The model is loosely based upon a published model by Engebretsen (2019) doi: 10.1371/journal.pcbi.1006879 .

For more information, look at vignette("commuter_model","spread")

Examples

spread::commuter(
  seiiar = spread::nor_seiiar_measles_oslo_2017_b2020,
  commuters = spread::nor_commuters_2017_b2020,
  r0 = 14,
  latent_period = 8,
  infectious_period = 5,
  asymptomatic_prob = 0,
  asymptomatic_relative_infectiousness = 0,
  days_simulation = 7 * 9,
  N = 1,
  simulations = 1,
  verbose = TRUE,
  seed = 4
)
#> Starting to add edges, printing every 1000 edge
#> 1000 2000 3000 4000 5000 6000 7000 8000 9000 10000 11000 12000 13000 14000 15000 16000 17000 18000 19000 20000 21000 22000 23000 Found 23898 edges
#> Running 1 simulations of 63 days
#> 
#> 
#> Finished all simulations
#>        sim_id   location_code week day is_6pm     S E I Ia      R incidence
#>     1:      1 municip_nor0301    1   1   TRUE 59764 3 9  0 606985         0
#>     2:      1 municip_nor0301    1   2   TRUE 59763 4 9  0 606985         0
#>     3:      1 municip_nor0301    1   3   TRUE 59762 5 7  0 606987         0
#>     4:      1 municip_nor0301    1   4   TRUE 59759 8 7  0 606987         0
#>     5:      1 municip_nor0301    1   5   TRUE 59758 9 5  0 606989         0
#>    ---                                                                     
#> 22424:      1 municip_nor5444    9  59   TRUE  1149 0 0  0   9050         0
#> 22425:      1 municip_nor5444    9  60   TRUE  1149 0 0  0   9050         0
#> 22426:      1 municip_nor5444    9  61   TRUE  1149 0 0  0   9050         0
#> 22427:      1 municip_nor5444    9  62   TRUE  1149 0 0  0   9050         0
#> 22428:      1 municip_nor5444    9  63   TRUE  1149 0 0  0   9050         0
#>           pop
#>     1: 666761
#>     2: 666761
#>     3: 666761
#>     4: 666761
#>     5: 666761
#>    ---       
#> 22424:  10199
#> 22425:  10199
#> 22426:  10199
#> 22427:  10199
#> 22428:  10199